migrate copyfile functions to dedicated module
